+++ docs/CHANGELOG.md
@@ -28,6 +28,7 @@ Note that genqlient is now tested from Go 1.20 through Go 1.22.
 
 ### New features:
 
+- The new `graphql.WithOperationNameParam` allows clients to be created that include `operationName` as a query parameter.
 - The new `optional: generic` allows using a generic type to represent optionality. See the [documentation](genqlient.yaml) for details.
 - For schemas with enum values that differ only in casing, it's now possible to disable smart-casing in genqlient.yaml; see the [documentation](genqlient.yaml) for `casing` for details.
 - Support .graphqls and .gql file extensions